09 ex00 almost done

+++ CPP09/ex00/BitcoinExchange.hpp
@@ -7,7 +7,7 @@
 class BitcoinExchange
 {
     private:
-        std::map<std::string, double> _data;
+        std::map<std::string, float> _data;
     public:
         BitcoinExchange();
         ~BitcoinExchange();
